api-server/requirements.txt
fastapi
uvicorn
requests
aiofiles
websocket-client
Pillow
python-multipart